Split out `match_single_binding`
